requirements.txt
click
prettytable
pyOpenSSL
